Fix issue inviting reviewers with no name
[BPHH-530]

User first/last name is required by the UI once the user has access to
the system. This was previously handled on the accept invitation screen
but that only covers the case where the user creates an account
manually. If the user uses BCeID to login, they cannot set their
first/last name and this results in bugs. For now, the easiest way to
fix this is to require the review manager to specify the first/last name
when inviting users.