Fix issue with building type select
[BPHH-809]

Fixes issue rendering building type once selected. This is due to an
HTML error where input elements cannot have children. Instead, we should
render the input as a flexbox to allow the value to be rendered as a
child.

Also fixes a UI issue where the building type select was pushing the
entire accordion section contents up, resulting in the first few fields
getting hidden. It's not clear why this happens as it doesn't happen for
other select fields which overflow the accordion panel as expected (e.g.
EPCTestingTargetTypeSelect). For now, we can set the popover to display
on top of the input.