Fix issue generating PDFs when deployed
[BPHH-1182]

Refactor the PDF generation code so it saves the PDF data to a temp file
and passes the filename rather than the JSON directly to the node
function. This fixes an error where the length of npm arguments is
restricted.

Also addresses an issue with PDF filename specificity. The previous
names are too generic as in the production environment, it is possible
that multiple application PDFs are generated at the same timei. The
name must be unique to the application to ensure the files are not
overridden before the PDFs have finished generating.