Fix issue clearing email on save
If the user clicks 'save' twice when onboarding, their email was getting
set to nil. This should only happen if the user already has a confirmed
email to allow them to switch back to this confirmed email. Otherwise,
the user gets into a state where their email is nil but they can still
get confirmed by clicking on the confirmation email link.