Fix error when omniauth fails due to invalid CSRF
[BPHH-1446]

If the user tries to login on a stale page with an old CSRF token, the
omniauth call fails. This was causing a 500 rather than failing
gracefully due to a comma in the `redirect_to` call. This resulted in the
`frontend_flash_message` getting passed as a second param to redirect_to`,
rather than getting passed to `login_path` as intended.