Fix issues (re)inviting users
[BPHH-1653]
[BPHH-1652]

Fixes issues inviting a user that already has a submitter account.

Previously, if a review staff member was invited but did not accept the
invite and instead created a submitter account, they could not always be
re-invited as a review staff. This occurred if the submitter was
retrieved first when inviting by email. In this scenario, the submitter
would be updated to have the same nickname as the existing review staff.
This triggered an error since the nickname field has a unique
constraint. This exposed the real error, which is that the submitter
should not be considered when retrieving the user to invite.

This also fixes a related issue where a submitter is invited as a review
staff for the first time. Instead of sending an invite to the review
staff as a new user, the submitter was automatically converted to a
review staff without needing to accept the invite.