Fix issue accepting regional RM invite
[BPHH-1736]

Fixes an issue where a logged in reviewer/review manger accepts a
regional RM invitation sent to a different email than the one they are
logged in with. Previously, this was upgrading the user to a regional RM
but since the regional RM was not invited with the proper jurisdictions,
the user only got assigned to their existing jurisdiction.